requirements.txt
fastapi
Pillow
uvicorn
stability-sdk